% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_imu.R
\name{animate_imu}
\alias{animate_imu}
\title{Animate movement of an inertial measurement unit}
\usage{
animate_imu(data, dt, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A \code{list} of numeric 4-vectors each of which a unit quaternion \link{w,x,y,z}}

\item{dt}{A numeric of time duration in milli-seconds}

\item{width, height}{width and height of animate_imu htmlwidget
specified in any valid \code{CSS} size unit}

\item{elementId}{\code{string} id as a valid \code{CSS} element id.}
}
\description{
Create interactive animated movements of an inertial measurement unit
}
