% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imu_object.R
\name{imu_object-update}
\alias{imu_object-update}
\alias{imu_proxy}
\alias{imu_send_data}
\title{Update a widget in shiny with an orientation}
\usage{
imu_proxy(id, session = shiny::getDefaultReactiveDomain())

imu_send_data(proxy, data)
}
\arguments{
\item{id}{HTML element id that houses the html widget}

\item{session}{shiny session - default to current session}

\item{proxy}{imu_proxy output to be used as input to imu_send_data}

\item{data}{a numeric unit 4-vector (quaternion) for the updated orientation}
}
\description{
Pair of functions used in conjunction to update a imu_object widget
}
