% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imuf-package.R
\docType{package}
\name{imuf-package}
\alias{imuf}
\alias{imuf-package}
\title{imuf: Estimate Orientation of a 6 dof IMU with a Complementary Filter}
\description{
An estimator of the orientation of an inertial measurement unit with a 3-axis accelerometer and a 3-axis gyroscope. You provide the accelerometer and gyroscope readings, time duration, the initial orientation, and a "gain" factor. We provide an estimate of the final orientation.
}
\author{
\strong{Maintainer}: Felix Chan \email{chanfelix@gmail.com}

}
\keyword{internal}
