% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walking_shin_1.R
\docType{data}
\name{walking_shin_1}
\alias{walking_shin_1}
\title{University of Mannheim Real World Activities}
\format{
\subsection{\code{walking_shin_1}}{

A data frame with 31,946 rows and 6 columns of accelerometer and gyroscope
measurements at 50 Hz
\describe{
\item{acc_x, acc_y, acc_z}{north, east and down acceleration in m/s^2}
\item{gyr_x, gyr_y, gyr_z}{angular velocity about north, east and down in rad/sec}
}
}
}
\source{
\url{https://www.uni-mannheim.de/dws/research/projects/activity-recognition/dataset/dataset-realworld/s1/}
}
\usage{
walking_shin_1
}
\description{
Accelerations and angular velocities captured by sensors of the mobile device
situated on the shin of subject 1 while the subject was walking for 10
minutes
}
\keyword{datasets}
