% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imuf-package.R
\docType{package}
\name{imuf-package}
\alias{imuf}
\alias{imuf-package}
\title{imuf: Estimate Orientation of a 6 Degree-of-Freedom Inertial Measurement Unit with a Complementary Filter}
\description{
An estimator of the orientation of an inertial measurement unit (IMU) with a 3-axis accelerometer and a 3-axis gyroscope. You provide 'imuf' the IMU's accelerometer and gyroscope readings, time duration, its initial orientation, and a "gain" factor. 'imuf' provides an estimate of the IMU's final orientation.
}
\author{
\strong{Maintainer}: Felix Chan \email{chanfelix@gmail.com} [copyright holder]

}
\keyword{internal}
