% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imuf-package.R
\docType{package}
\name{imuf-package}
\alias{imuf}
\alias{imuf-package}
\title{imuf: Estimate Orientation of an Inertial Measurement Unit}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Estimate the orientation of an inertial measurement unit (IMU) with a 3-axis accelerometer and a 3-axis gyroscope using a complementary filter. 'imuf' takes an IMU's accelerometer and gyroscope readings, time duration, its initial orientation, and a gain factor as inputs, and returns an estimate of the IMU's final orientation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gitboosting/imuf}
  \item \url{https://gitboosting.github.io/imuf/}
  \item Report bugs at \url{https://github.com/gitboosting/imuf/issues}
}

}
\author{
\strong{Maintainer}: Felix Chan \email{chanfelix@gmail.com} [copyright holder]

}
\keyword{internal}
