% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rotV}
\alias{rotV}
\title{Rotate a 3-vector by a quaternion}
\usage{
rotV(quat, vin)
}
\arguments{
\item{quat}{A numeric unit 4-vector (w,x,y,z) for a rotation quaternion}

\item{vin}{A numeric 3-vector to be rotated by quat}
}
\value{
A numeric 3-vector after the rotation
}
\description{
\code{rotV()} rotates a 3-vector by a quaternion expressed as a unit 4-vector in
(w,x,y,z) convention
}
\examples{
q <- c(cos(pi/4), sin(pi/4), 0, 0)
vin <- c(0, 1, 0)
rotV(q, vin)

}
